Add Rust CI and align core Cashu helpers with upstream cashu crate
Agent-Logs-Url: https://github.com/Amperstrand/micronuts/sessions/5e27c623-5df5-4829-9d42-c91c55d4f9e2

+++ cashu-core-lite/src/crypto.rs
@@ -27,13 +27,17 @@ impl std::error::Error for HashToCurveError {}
 ///
 /// where `msg_hash = SHA256("Secp256k1_HashToCurve_Cashu_" || x)`,
 /// counter is `u32` little-endian from 0 to `u16::MAX - 1`.
+///
+/// This matches the upstream Cashu/CDK reference implementation, which hashes a
+/// 4-byte little-endian counter (not 2-byte) while still limiting the search to
+/// the first `u16::MAX` values.
 pub fn hash_to_curve(message: &[u8]) -> Result<PublicKey, HashToCurveError> {
     let msg_hash = Sha256::new()
         .chain_update(DOMAIN_SEPARATOR)
         .chain_update(message)
         .finalize();
 
-    for counter in 0..u16::MAX {
+    for counter in 0..u16::MAX as u32 {
         let candidate = Sha256::new()
             .chain_update(msg_hash)
             .chain_update(counter.to_le_bytes())
